refactor(ship): adresse la review (throw, success embed, camelCase history, helper dans /utils)
