+++ apps/bot/src/domains/ship/utils/index.ts
@@ -1,3 +1,9 @@
 export { buildShipModuleUpgradePreview } from './build-ship-module-upgrade-preview.js';
 export { parseShipModuleKey } from './parse-ship-module-key.js';
-export { getShipModuleBerryCost, getShipModuleLevel, getShipModuleResourceCosts, isShipModuleMaxLevel } from './ship-module-helpers.js';
+export {
+  getMaxHpForHullLevel,
+  getShipModuleBerryCost,
+  getShipModuleLevel,
+  getShipModuleResourceCosts,
+  isShipModuleMaxLevel,
+} from './ship-module-helpers.js';
