+++ apps/bot/src/domains/ship/utils/ship-module-helpers.ts
@@ -8,6 +8,13 @@ export function getShipModuleLevel(ship: Ship, moduleKey: ShipModuleKey): number
   return ship[SHIP_MODULE_LEVEL_COLUMNS[moduleKey]];
 }
 
+/** HP max d'un navire = valeur de la coque au niveau courant. Source unique pour la barre HP, le drift et l'init d'un template. */
+export function getMaxHpForHullLevel(hullLevel: number): number {
+  const maxHp = SHIP_MODULES.hull.valueByLevel[hullLevel - 1];
+  if (maxHp === undefined) throw new Error(`Niveau de coque invalide : ${hullLevel}`);
+  return maxHp;
+}
+
 export function isShipModuleMaxLevel(moduleKey: ShipModuleKey, level: number): boolean {
   return level >= SHIP_MODULES[moduleKey].valueByLevel.length;
 }
